Move instrumentation testing
